@@ -1,24 +1,33 @@
 {
-  "name": "tango playlist-app",
+  "name": "tanda-app",
   "version": "0.0.1",
-  "description": "A little tango playlist maker",
+  "description": "A little tanda maker",
+  "type": "module",
   "main": "server.js",
   "scripts": {
     "start": "node server.js"
   },
-  "dependencies": {
-    "express": "^4.17.1",
-    "winston": "^3.2.1",
-    "express-handlebars": "^3.1.0",
-    "body-parser": "^1.19.0",
-    "uuid": "^3.4.0",
-    "lowdb": "^0.14.0",
-    "lodash": "^4.17.15"
+
+"dependencies": {
+  "body-parser": "^1.20.2",
+  "cloudinary": "^1.34.0",
+  "cookie-parser": "^1.4.6",
+  "express": "^4.18.2",
+  "express-fileupload": "^1.4.0",
+  "express-handlebars": "^7.0.1",
+  "lodash": "^4.17.21",
+  "lowdb": "^5.1.0",
+  "uuid": "^9.0.0",
+    "winston": "^3.6.0"
+
+},
+
+
+"engines": {
+    "node": "16.x"
   },
-  "engines": {
-    "node": "10.x"
-  },
-  "keywords": [
+  
+"keywords": [
     "node",
     "glitch",
     "express"
