@@ -1,4 +1,4 @@
-const winston = require('winston');
+import winston from 'winston';
 
 const logger = new (winston.createLogger)({
   transports: [new (winston.transports.Console)({ json: false })],
@@ -10,4 +10,4 @@ if (process.env.LEVEL) {
   logger.level = process.env.LEVEL;
 }
 
-module.exports = logger;
+export default logger;
