@@ -1,5 +1,3 @@
-/* If you're feeling fancy you can add interactivity 
-    to your site with Javascript */
 
 // prints "hi" in the browser's dev tools console
 console.log("Hello from the tango project!");
@@ -7,38 +5,39 @@ console.log("Hello from the tango project!");
 const greenbtn = document.querySelector(".green");
 
 greenbtn &&
-  greenbtn.addEventListener("click", () => alert("Thanks! You're okay too"));
+  greenbtn.addEventListener("click", () => alert("Thank you for your appreciation!"));
 
 const bluebtn = document.querySelector(".blue");
 
 bluebtn &&
   bluebtn.addEventListener("click", () => {
-    let readMoreDiv = document.querySelector("#readmore");
-    readMoreDiv.style.color = "green";
-    if (readMoreDiv.style.display === "block") {
-      readMoreDiv.style.display = "none";
+    let infoDiv = document.querySelector("#info");
+    //infoDiv.style.color = "green";
+    if (infoDiv.style.display === "block") {
+      infoDiv.style.display = "none";
     } else {
-      readMoreDiv.style.display = "block";
+      infoDiv.style.display = "block";
     }
   });
 
 const redbtn = document.querySelector(".red");
 
-const welcomeUserDiv = document.querySelector("#welcomeuser");
+const introductionDiv = document.querySelector("#introduction");
 
 redbtn &&
   redbtn.addEventListener("click", () => {
-    let username = prompt("What's your name?");
-    welcomeUserDiv.style.display = "block";
-    document.querySelector("#welcomeuser").innerHTML = `<p> Hello, ${username}, 
+    let intro = prompt("What is your name tanguero?");
+    //introductionDiv.style.color = "darkred";
+    introductionDiv.style.display = "block";
+    document.querySelector("#introduction").innerHTML = `<p> Hello, ${intro}, 
     I can't wait to hearing your tango playlists! Click this message to close it.</p>`;
-    welcomeUserDiv.style.cursor = "pointer";
+    introductionDiv.style.cursor = "pointer";
   });
 
-welcomeUserDiv &&
-  welcomeUserDiv.addEventListener("click", (evt) => {
+introductionDiv &&
+  introductionDiv.addEventListener("click", (evt) => {
    // evt.currentTarget.style.display = "none";
-    welcomeUserDiv.style.display = "none";
+    introductionDiv.style.display = "none";
   });
 
 const ratebtn  = document.querySelector("#rateit");
@@ -47,11 +46,13 @@ ratebtn &&
   ratebtn.addEventListener("click", () => {
    let userRating = parseInt(prompt("Rate this Tango Collection (from 1 to 5 stars)"));
   if (userRating>5 || userRating<1 || isNaN(userRating)){
-    alert("Oops, try again with a number between 1 and 5!");
+    alert("One more time, with a number between 1 and 5!");
   }
   else{
 
     document.querySelector("#rating").innerHTML = "You gave a rating of: ";
+  
+   // document.getElementById("#theRating").innerHTML = userRating;
     for (let i=0; i < userRating; i++){
         document.querySelector("#rating").innerHTML +="<i class='yellow star icon'></i>";
     }
@@ -60,4 +61,4 @@ ratebtn &&
 
 $(".delmelody").click(() => confirm('Really delete this tango?'))
 
-$(".deltangoPlaylist").click(() => confirm('Really delete this tango playlist?'))
+$(".deltanda").click(() => confirm('Really delete this tango playlist?'))
